+++ app/diagram-pack-generator/page.tsx
@@ -1,0 +1,5 @@
+import DiagramPackGenerator from "../../components/DiagramPackGenerator";
+
+export default function DiagramPackGeneratorPage() {
+  return <DiagramPackGenerator />;
+}